unit: deserialize keeps all semantic data
The existing unit.Deserialize() function does not contain all the
semantic data in a unit file. Specifically it does not fully
support multiple sections. It does correctly parse the sections,
but the `[]*UnitOption` data structure does not tell the caller,
in the case of a unit file with duplicate section names, *which*
section the UnitOptions are from.

This may matter.

For example, a `/etc/systemd/network/*.network` file can have multiple
`[Route]` entries. Just returning an array of `Section, Name, Value`
tuples does not tell the caller which `[Route]` a given `Name, Value`
pair belongs to.

Ex:
```
[Route]
Gateway=10.0.100.1
Destination=10.0.0.1/24

[Route]
Gateway=10.0.100.2
Destination=10.0.2.1/24
```

This unit file will generate the following from `unit.Deserialize()`

```
{Section: "Route", Name: "Gateway", Value: "10.0.100.1"}
{Section: "Route", Name: "Destination", Value: "10.0.0.1/24"}
{Section: "Route", Name: "Gateway", Value: "10.0.100.2"}
{Section: "Route", Name: "Destination", Value: "10.0.2.1/24"}
```

With just this information, it is not possbile to know which `Gateway`
goes to which `Destination`. Order maybe, but what if there were not two
Gateways? Which `Destination` would go with the `Gateway`?

This patch introduces a new Deserialize function that returns the
following data structures:

```
// UnitEntry is a single line entry in a Unit file.
type UnitEntry struct {
	Name  string
	Value string
}

// UnitSection is a section in a Unit file. The section name
// and a list of entries in that section.
type UnitSection struct {
	Section string
	Entries []*UnitEntry
}
```

The new DeserializeSection() signature is:

```
func DeserializeSections(f io.Reader) ([]*UnitSection, error)
```

And the output from it given the example unit file above is:

```
{Section: Route{Name: Gateway, Value: 10.0.100.1}{Name: Destination, Value: 10.0.0.1/24}}
{Section: Route{Name: Gateway, Value: 10.0.100.2}{Name: Destination, Value: 10.0.2.1/24}}
```

The `Name`, `Value` pairs are directly associated with a specific
section. There is no ambiguity.

There are unit tests added to `unit/deserialize_test.go` that do basic
tests on `DeserializeSections()`. All exising unit tests in that file
are untouched and pass.